Added PWM Compatible code
This version of the tank Arduino code is capable of speed adjustments, the tank can go left and right while both tracks are in motion, also throtle will be adjustable for the next version, which will be uploaded in an hour I guess. I'm making these changes to control the tank with an XBOX 360 controller.